HARP-9151: Remove ExtendedTileInfo decoding from OMV and GeoJson.
createTileInfo and gatherRoadSegments OMV options are now deprecated.
