+++ @here/harp-datasource-protocol/lib/DecodedTile.ts
@@ -14,7 +14,15 @@ import {
     Vector3Like,
     webMercatorProjection
 } from "@here/harp-geoutils";
-import { IndexedTechnique } from "./Techniques";
+import { Env } from "./Expr";
+import { AttrEvaluationContext, evaluateTechniqueAttr } from "./TechniqueAttr";
+import {
+    IndexedTechnique,
+    isLineMarkerTechnique,
+    isPoiTechnique,
+    isTextTechnique,
+    Technique
+} from "./Techniques";
 import { TileInfo } from "./TileInfo";
 
 /**
@@ -325,4 +333,92 @@ export function getFeatureId(attributeMap: AttributeMap | undefined): number {
     }
 
     return 0;
+}
+
+/**
+ * Determine the name of (OMV) feature. It implements the special handling required
+ * to determine the text content of a feature from its tags, which are passed in as the `env`.
+ *
+ * @param env Environment containing the tags from the (OMV) feature.
+ * @param useAbbreviation `true` to use the abbreviation if available.
+ * @param useIsoCode `true` to use the tag "iso_code".
+ * @param languages List of languages to use, for example: Specify "en" to use the tag "name_en"
+ *                  as the text of the string. Order reflects priority.
+ */
+export function getFeatureName(
+    env: Env,
+    useAbbreviation?: boolean,
+    useIsoCode?: boolean,
+    languages?: string[]
+): string | undefined {
+    let name;
+    if (useAbbreviation) {
+        const abbreviation = env.lookup(`name:short`);
+        if (typeof abbreviation === "string" && abbreviation.length > 0) {
+            return abbreviation;
+        }
+    }
+    if (useIsoCode) {
+        const isoCode = env.lookup(`iso_code`);
+        if (typeof isoCode === "string" && isoCode.length > 0) {
+            return isoCode;
+        }
+    }
+    if (languages !== undefined) {
+        for (const lang of languages) {
+            name = env.lookup(`name:${lang}`) || env.lookup(`name_${lang}`);
+            if (typeof name === "string" && name.length > 0) {
+                return name;
+            }
+        }
+    }
+    name = env.lookup("name");
+    if (typeof name === "string") {
+        return name;
+    }
+    return undefined;
+}
+
+/**
+ * Determine the text string of the map feature. It implements the special handling required
+ * to determine the text content of a feature from its tags, which are passed in as the `env`.
+ *
+ * @param feature Feature, including properties from the (OMV) feature.
+ * @param technique technique defining how text should be created from feature
+ * @param languages List of languages to use, for example: Specify "en" to use the tag "name_en"
+ *                  as the text of the string. Order reflects priority.
+ */
+export function getFeatureText(
+    context: Env | AttrEvaluationContext,
+    technique: Technique,
+    languages?: string[]
+): string | undefined {
+    let useAbbreviation: boolean | undefined;
+    let useIsoCode: boolean | undefined;
+    const env = context instanceof Env ? context : context.env;
+    if (
+        isTextTechnique(technique) ||
+        isPoiTechnique(technique) ||
+        isLineMarkerTechnique(technique)
+    ) {
+        if (technique.text !== undefined) {
+            return evaluateTechniqueAttr(context, technique.text);
+        }
+        // tslint:disable-next-line: deprecation
+        if (technique.label !== undefined) {
+            // tslint:disable-next-line: deprecation
+            const attributeName = evaluateTechniqueAttr(context, technique.label);
+            if (typeof attributeName !== "string") {
+                return undefined;
+            }
+            const name = env.lookup(attributeName);
+            return typeof name === "string" ? name : undefined;
+        }
+        // tslint:disable-next-line: deprecation
+        useAbbreviation = technique.useAbbreviation;
+        // tslint:disable-next-line: deprecation
+        useIsoCode = technique.useIsoCode;
+    }
+
+    return getFeatureName(env, useAbbreviation, useIsoCode, languages);
 }
